Add context to page and menu title "Activity Log" strings
Adding this context will allow to separate Plugin Name from page and menu titles.
The Plugin Name shouldn't be translated, keeping the original is the best practice, as is the name it is known and it have a reputation built across languages.
The page and menu titles can/should be translated to improve usability.
This is a common problem where the plugin name is exactly it's function and menu title, also very common in simple and straight forward plugin names :)