Fix callings of $wpdb->prepare due to core documentation
1. Fixed table name in placeholders. Lines like this is incorrect:
> SELECT * FROM `%1$s`
You should not swap out the tablename. If you do, the table name will be wrapped in quotes and it will trigger a SQL error in some (depend of server settings) cases.

`prepare()` is meant to operate on user supplied data - data that could be from questionable sourced with malicious intent. You do not need to swap out data, like your table name, that is not from a questionable source. So I fixed this.

2. Wordpress `prepare()` method haven't support for numerated placeholders anymore. Early, was possible call `prepare()` with placeholders like `%1$s`, but due to [this issue](https://medium.com/websec/wordpress-sqli-bbb2afcc8e94) and starts from wp 4.8.x and higher, its impossible. Ticket and discussion about numerated placeholders is here: https://core.trac.wordpress.org/ticket/41925

I replaced all placeholders with position specifier to common placeholders which have a full support.