Fix WordPress 4.8.2 & 4.7.6 compatibility issues
Fix: $wpdb->prepare() issues with numerated placeholders in WordPress 4.8.2 & 4.7.6.
Fix: A couple of typos.
Change: Usage of numerated placeholders in `printf()` when not needed and for consistancy.
Change: Escaping placeholders with quotes when not needed as it's handled by `prepare()`.
Change: Set default order to DESC and the default order by as `hist_time`.