Fix: Request-URI Too Long error
After filtering the log a few times a *Request-URI Too Long* error will be thrown due to the URL length exceeding the capacity limit for most servers. This is caused by the `_wpnonce` and `_wp_http_referer` parameters being added to the URI multiple times. `wp_none_field()` function isn't currently being used for anything in this plugin and as such can be removed, fixing the above issue.